requirements.txt
asgiref==3.3.4
Django==3.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
PyJWT==2.1.0
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1
